refactor: move SAM-debug stuff to sam/debugger/
