+++ src/shared/sam/debugger/awsSamDebugger.ts
@@ -8,9 +8,9 @@ import * as nls from 'vscode-nls'
 import { getCodeRoot, getHandlerName, getTemplateResource } from '../../../lambda/local/debugConfiguration'
 import { getDefaultRuntime, getFamily, getRuntimeFamily, RuntimeFamily } from '../../../lambda/models/samLambdaRuntime'
 import { CloudFormationTemplateRegistry, getResourcesFromTemplateDatum } from '../../cloudformation/templateRegistry'
-import * as csharpDebug from '../../codelens/csharpCodeLensProvider'
-import * as pythonDebug from '../../codelens/pythonCodeLensProvider'
-import * as tsDebug from '../../codelens/typescriptCodeLensProvider'
+import * as csharpDebug from './csharpSamDebug'
+import * as pythonDebug from './pythonSamDebug'
+import * as tsDebug from './typescriptSamDebug'
 import { ExtContext } from '../../extensions'
 import { isInDirectory } from '../../filesystemUtilities'
 import { getLogger } from '../../logger'
