+++ src/shared/sam/localLambdaRunner.ts
@@ -9,16 +9,16 @@ import * as tcpPortUsed from 'tcp-port-used'
 import * as vscode from 'vscode'
 import { detectLocalLambdas, LocalLambda } from '../../lambda/local/detectLocalLambdas'
 import { getFamily, RuntimeFamily } from '../../lambda/models/samLambdaRuntime'
-import * as pathutil from '../../shared/utilities/pathUtils'
+import * as pathutil from '../utilities/pathUtils'
 import { CloudFormation } from '../cloudformation/cloudformation'
 import { ExtContext } from '../extensions'
 import { makeTemporaryToolkitFolder } from '../filesystemUtilities'
 import { getLogger } from '../logger'
-import { DefaultValidatingSamCliProcessInvoker } from '../sam/cli/defaultValidatingSamCliProcessInvoker'
-import { SamCliBuildInvocation, SamCliBuildInvocationArguments } from '../sam/cli/samCliBuild'
-import { SamCliProcessInvoker } from '../sam/cli/samCliInvokerUtils'
-import { SamCliLocalInvokeInvocation, SamCliLocalInvokeInvocationArguments } from '../sam/cli/samCliLocalInvoke'
-import { SamLaunchRequestArgs } from '../sam/debugger/samDebugSession'
+import { DefaultValidatingSamCliProcessInvoker } from './cli/defaultValidatingSamCliProcessInvoker'
+import { SamCliBuildInvocation, SamCliBuildInvocationArguments } from './cli/samCliBuild'
+import { SamCliProcessInvoker } from './cli/samCliInvokerUtils'
+import { SamCliLocalInvokeInvocation, SamCliLocalInvokeInvocationArguments } from './cli/samCliLocalInvoke'
+import { SamLaunchRequestArgs } from './debugger/samDebugSession'
 import { SettingsConfiguration } from '../settingsConfiguration'
 import { recordLambdaInvokeLocal, recordSamAttachDebugger, Result, Runtime } from '../telemetry/telemetry'
 import { TelemetryService } from '../telemetry/telemetryService'
