+++ src/test/shared/codelens/localLambdaRunner.test.ts
@@ -8,7 +8,7 @@ import * as del from 'del'
 import { readdir } from 'fs-extra'
 import * as path from 'path'
 import * as vscode from 'vscode'
-import * as localLambdaRunner from '../../../shared/codelens/localLambdaRunner'
+import * as localLambdaRunner from '../../../shared/sam/localLambdaRunner'
 import * as fsUtils from '../../../shared/filesystemUtilities'
 import { ChildProcessResult } from '../../../shared/utilities/childProcess'
 import { FakeChannelLogger } from '../fakeChannelLogger'
